requirements.txt
python-dotenv==1.0.0
requests==2.31.0
tweepy==4.14.0
pytrends==4.9.2
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.26.3
scikit-learn==1.3.2
pytest==7.4.3
python-dateutil==2.8.2
newsapi-python==0.2.7
transformers==4.36.0
torch==2.1.0
statsmodels==0.14.1
prophet==1.1.5
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
